First major work on 6.0.0
