+++ build.gradle
@@ -11,6 +11,28 @@ plugins{
 
 def ver = new Version(major: 5, minor: 2, revision: 3)
 
+allprojects {
+    apply plugin: 'com.jfrog.bintray'
+    apply plugin: 'maven-publish'
+    apply plugin: 'java-library'
+    apply plugin: 'com.github.johnrengelman.shadow'
+    
+    group = "org.botblock"
+    version = "$ver"
+    
+    repositories {
+        mavenCentral()
+        jcenter()
+    }
+    
+    dependencies {
+        api group: 'com.squareup.okhttp3', name: 'okhttp', version: '4.7.2'
+        api group: 'org.json', name: 'json', version: '20200518'
+        api group: 'com.github.ben-manes.caffeine', name: 'caffeine', version: '2.8.0'
+        api group: 'com.google.code.findbugs', name: 'jsr305', version: '3.0.2'
+    }
+}
+
 group = "org.botblock"
 version = "$ver"
