+++ jda-module/build.gradle
@@ -1,0 +1,14 @@
+plugins {
+    id 'java'
+}
+
+group 'org.botblock'
+version '5.2.3'
+
+dependencies {
+    api(group: 'net.dv8tion', name: 'JDA', version: '4.2.0_172'){
+        exclude(module: 'opus-java')
+    }
+    implementation project(":core")
+    implementation project(":request-module")
+}
