+++ request-module/build.gradle
@@ -1,0 +1,7 @@
+plugins {
+    id 'java'
+}
+
+dependencies {
+    implementation project(":core")
+}
