+++ settings.gradle
@@ -1,2 +1,5 @@
 rootProject.name = 'JavaBotBlockAPI'
+include 'core'
+include 'jda-module'
+include 'request-module'
